Refuse to read installable control data in national encodings.
Finally processes control files consistently between sources and
installable packages. Implements the Unicode release goal [1] with
respect to our primary package control files.

Some thought was given to issuing a tag instead of a hard program
error, but the resulting complexity did not outweigh the expected
benefits of this change.

[1] https://wiki.debian.org/ReleaseGoals/utf-8