Prefer spaces over commas as field separators in profiles.
Commas are not needed, and spaces are more convenient.  None of
Lintian's identifiers contain spaces.  All other parts of Lintian like
tag and test specifications already use spaces.

Commas are still accepted but, from now on, they generate a warning.

Please also note that newlines are now trimmed transparently. That
allows a clear multi-line formatting style everywhere.

It is easy on the eyes and much better for diffs. In Git, it reduces
the complexity of merge conflicts.  For example, the kali.main profile
now reads:

    Disable-Tags:
     bugs-field-does-not-refer-to-debian-infrastructure
     no-nmu-in-changelog
     source-nmu-has-incorrect-version-number

Instead of previously:

    Disable-Tags: bugs-field-does-not-refer-to-debian-infrastructure,
     no-nmu-in-changelog, source-nmu-has-incorrect-version-number