In harness test, continue to allow commas as field separators for profiles.
All profiles we ship were converted. There would be no occasion to
allow commas in our harness test except our test checks all profiles
that are present.

The test works fine on a development source tree but not when testing
an installed version, as autopkgtest does. Lintian depends on dpkg,
which installs its own profile. That profile was not yet converted.
A single tag name there is misparsed to include a comma at the end:

    Failed test 'Tag package-uses-vendor-specific-patch-series, exists
    in profile /usr/share/lintian/profiles/dpkg/main.profile'

The error showed up only in Salsa CI. This commit should fix it.

A better course of action might be to test only our profiles, if
possible, or perhaps to test none at all.

Gbp-Dch: ignore