Anchor some negative regex assertions for architectures in the rear, and all in front. (See: !300)
Jiří Paleček made some good points about these regular expressions in
the MR noted above, but I do not agree with the suggested approach to
solving the problem.

I believe the regular expressions at issue here require anchors in the
rear so the negative (?!GNU) assertions hold for the full remainder of
each string, and not just for a part of them.

The anchors are added here.

I came up with what I think is a better solution when looking at the
example at https://regex101.com/r/14KozX/1 that Jiří so kindly
provided.

Jiří's wording also showed the way by stating that the expression
matches (his quotes) "a line which contains a word starting with
GNU/Linux, or any other word not starting with GNU". Matching words
is not enough. We wish to match the entire line, and anchoring does
exactly that.

Thanks to Jiří Paleček for the thoughtful suggestion, and for the
extraordinary patience before the merge request was resolved!