Treat NMU edge cases for non-consecutive-debian-revision. (Closes: #942013)
The bug report was confirmed by several parties. It even offers some
history because it was once reopened. Unfortunately, none of the false
positives mentioned in it seemed to work anymore.

In theory, this commit should conclusively resolve the bug, yet right
now the only confirmation for that claim is the false positive fixed
in the test suite. Please reopen the bug and amend it, if needed.

This effort was prompted by Mattia Rizzolo on IRC, who appealed
repeatedly for attention to this bug over the course of several
months.  Thank you for your patience, Mattia!