Always print full path to patch files. (Closes: #980247)
Makes the context for tags related to patches more consistent. Patch
files are now always identified via their full path in the source
package.

As the filer noted, showing the full path may ease packaging for less
experienced maintainers. They are also more likely to rely on Lintian
for guidance.

More experienced users may prefer the short path. We are instead
making tag names shorter, and the context longer. For tags related to
patches, it means that paths are getting longer.

Working with tag names often, we believe that shorter tag names are
easier to remember. Accordingly, we try to be precise with tag names
only when needed, for instance when tags distinguish closely related
topics.  Whenever possible, we provide any necessary precision in the
tag context. It makes tags more versatile.

This philosophy applies to many tags. For example, a relatively large
subgroup of names contains the word "debian". (Tags for changelogs or
watch files are good examples.) The word usually indicates that a
source file is located inside the ./debian directory. We perceive such
a tag naming scheme as cumbersome.

Many parts of our packaging infrastructure are named after Debian. The
moniker is generally overloaded, and often ambiguous. It is much
clearer to provide the full path of a file.

As a side note, we also no longer preface tag names with the name of
the check that issues them (although such a scheme still exists with
name-spaced tags; see continuous-integration.)

We sincerely hope that this solution is acceptable even though the
issue was resolved against the filer's stated preference.

Thank you to Alex Beckert for his keen eye, and also for remaining one
of our most active bug reporters!