Tolerate manual pages for subcommands when main command is documented. (Closes: #947258)
Prevents the tag spare-manual-page from being isseud for subcommands
like 'git add' in packages like notmuch or libreswan. Works only when
the main command has its own manual page.

The use of the underscore in libreswan's file names required special
casing because underscores are considered word characters.

Thanks to dkg for the interest in this bug, as well as the repeated
pointers and the necessary patience until now!