Flag Rubygem homepages in source control files. (Closes: #981935)
The condition here received its own tag. Eventually, it could be
desirable combine the tag with similar warnings for other languages or
packaging systems, but Lintian's rule of issuing each tag only in one
check would drop the logic from the Ruby language, where it arguably
belongs.

On the upside, the custom tag allows for a generous explanation with
Rubygem-specific references. Users will find it easier to understand.

Thanks to Paul Wise for bringing this matter to our attention!