Do not pretend to show hashbang in hint context for unusual interpreters. (Closes: #982630)
For scripts using 'env', the hint context was confusing because it did
not match the actual hashbang. We now list only the offending
interpreter without a hashbang (or 'env', if it was used).

The bug report makes other important points. They are adressed below:

> Either way, the warning message shouldn't omit /usr/bin/env if it's
> present in the file in question, as the mismatch between the warning and
> the file is a distraction from the intent of the message.

This check traditionally disregarded the use of 'env' in scripts (but
not in maintainer scripts). It would be inappropriate to show the full
hashbang.The tag relates only to the interpreter portion.

> If lintian is in fact trying to complain about "python" vs "python3" (if
> we're trying to ensure that we don't have anything pointing to
> "python"), that is a sufficiently important, specific situation that it
> should offer a dedicated tag for it.

I agree with your sentiment. Unfortunately, the interpreter tags are
already differentiated, i.e. in your case 'examples'.  The wisdom of
the existing distinctions is not obvious and seems inferior to the
issue you raised, but implementing your suggestion would require many
new tags, which we are unlikely to embrace.

We would, however, entertain a request to simplify the existing
interpreter tags in order to flag the unversioned Python call in a
better way.  Please file a bug if you believe that it is a worthwhile
trade-off for users.

> /usr/bin/python is supposed to be used for things that are both py3-
> and py2-compatible (which hopefully these examples all are).

Based on the information we have, the unversioned /usr/bin/python is
going away in the upcoming 'bullseye' release. The corresponding
changes were made in response to this merge request

    https://salsa.debian.org/lintian/lintian/-/merge_requests/334

in this and surrounding commits:

    https://salsa.debian.org/lintian/lintian/-/commit/69d52d7b4c9dbf391b23dced6e6de920905f54ac

Please file a bug for the unversioned Python interpreter to be
recognized as valid, if the information we have about the 'bullseye'
release is incorrect.

Thank you to dkg for bringing these matters to our attention!