Drop 'use autodie' everywhere; handle errors explicitly.
The 'autodie' pragma breaks many expectations for error handling and
fails to emulate parser quirks in functions like system and exec. Much
of that behavior is documented in the last paragraph here, and also in
the "Gotchas" section below:

    https://perldoc.perl.org/autodie#FUNCTION-SPECIFIC-NOTES

Thanks to Grinnz from IRC:#perl-help for these pointers!

Gbp-Dch: ignore