Assume that all interactive terminals are hyperlink capable but change default to off.
The implementation of OSC 8 hyperlinks [1] has come a long way. Many
programs enable hyperlinks for all terminals and use a blacklist for
terminals that are known not to support them.

GCC is reported to do that especially well. Lintian should ultimately
implement something similar. For the time being, we merely assume that
all interactive terminals are hyperlink capable, but no blacklist is
implemented at this time.

To avoid stray characters in terminals that not support the feature,
the display of hyperlinks was disabled by default. Users may wish to
add 'hyperlinks = on' to their Lintian configuration.

[1] https://gist.github.com/egmontkob/eb114294efbcd5adb1944c9f3cb5feda