Centralize operation of xargs in preparation for implementation in Perl.
A Perl implementation of xargs would make it possible to associate the
input for each invocation of the underlying command to the resulting
output.

It is helpful for programs like 'readelf' that produce different
output when called with multiple arguments. The number of arguments is
easily available, whicn helps with parsing.

Perhaps more significantly, it also allows the use of parallel
execution from within Perl without the locking described in the 'xarg'
manual page.