Flag systemd temporary files placed in /var/run. (Closes: #984678)
Like the related tag systemd-service-file-refers-to-var-run this tag
is issued at the info level. (The bug report used an old name.)

The list from the codesearch.d.o link in the report flags package
sources, but not all of those files in source packages are actually
being installed. The new tag looks at installed files.

We were able to validate the tag with the following packages (but not,
for example, with celery, sagan, samba or wesnoth-14):

$ bin/lintian /mirror/debian/pool/main/k/krb5/krb5-otp_1.18.3-4_amd64.deb
I: krb5-otp: systemd-tmpfile-in-var-run usr/lib/tmpfiles.d/krb5-otp.conf

$ bin/lintian /mirror/debian/pool/main/v/vrfydmn/vrfydmn_0.11.0-1_all.deb
I: vrfydmn: systemd-tmpfile-in-var-run usr/lib/tmpfiles.d/vrfydmn.conf

$ bin/lintian /mirror/debian/pool/main/v/vsftpd/vsftpd_3.0.3-13_amd64.deb
I: vsftpd: systemd-tmpfile-in-var-run usr/lib/tmpfiles.d/vsftpd.conf

$ bin/lintian /mirror/debian/pool/main/x/xrootd/xrootd-server_5.0.3-4_amd64.deb
I: xrootd-server: systemd-tmpfile-in-var-run usr/lib/tmpfiles.d/xrootd.conf

$ bin/lintian /mirror/debian/pool/main/z/zoneminder/zoneminder_1.34.23-1_amd64.deb
I: zoneminder: systemd-tmpfile-in-var-run usr/lib/tmpfiles.d/zoneminder.conf

As a side note, the related tag for tmpfile declarations installed in
the wrong location was renamed to misplaced-systemd-tmpfiles.  The
correct location is described in that tag's description.  It is part
of our effort to make tag names less specific when the details is not
needed to distinguish tags from another.

This commit regenerates the profiles.

Thanks to Michael Biebl for this clearly written and well-researched
feature request!