Read error message from Config::Tiny via the advertised mechanism. (Closes: #985064)
The POD for the module [1] explains how to read the error value, but
that procedure was not followed. This implements it correctly.

Lintian now produces the more meaningful error it was supposed to
generate in the first place:

$ bin/lintian --cfg /nonexistent /mirror/debian/pool/main/n/nyancat/nyancat_1.5.2-0.1_i386.deb
syntax error in configuration file /nonexistent: Failed to open file '/nonexistent' for reading: No such file or directory

Thanks to Jakub Wilk for bringing the matter to our attention!

[1] https://metacpan.org/pod/Config::Tiny#errstr()