Use semantic versioning for the output of '--version'.
Starting with version 2.6.0, Lintian adopted [1] a semantic versioning
scheme, but between releases we used the output of 'git describe'.
That format was incompatible.

From now on, we use the unreleased commit count as the "patch level"
which under our current scheme is zero for released versions. The
commit reference is kept as what SemVer calls "build metadata" for
easier navigation in git.

The versioning issue came up in Lintian's new web site. The packaging
hints displayed there are sometimes produced with unreleased versions.

At the same time, the Postgres database holds those hints and the tag
descriptions for different Lintian versions. The 'text' data type is
insufficient for ordering those columns. We plan to use a database
extension with a 'semver' data type [2] for that purpose.

We could probably use Debian's own 'debversion' data type [3] but that
module is presumably less well-tested. It furthermore suffers from
performance problems described in the open Bug#935637 (even though
those may not matter for the narrower version scheme Lintian adopted).

[1] see "Misc", https://salsa.debian.org/lintian/lintian/-/tags/2.6.0
[2] https://pgxn.org/dist/semver/doc/semver.html
[3] https://tracker.debian.org/pkg/postgresql-debversion