Disregard a source's current Standard-Version when looking for newer policy releases.
The 'before' expression is a bit confusing because it disgards
everthing that meets the condition. The new condition causes the
current compliance version to be excluded. The fourth digit is now
being ignored as a non-normative policy change, just as it should.

This bug produced false positives for a wide range of packages,
including for Lintian itself.

$ bin/lintian ../lintian_2.104.0+git.dsc
I: lintian source: out-of-date-standards-version 4.5.1 (released 2020-11-17) (current is 4.5.1.1)