Omit Git hash from all Lintian version strings.
For unreleased versions, we used to store the short commit hash from
'git describe' in the database, but stripped it for display purposes
in the web site. That ended up being more complex than necessary,
because the hash was never needed.

Drops the hash from all version strings. It will simplify the code in
the web site and the SQL queries significantly.