Build test packages automatically when running tests, even for partial tests. (Closes: #927476, #935907)
From now on, private/runtests automatically builds all test packages
needed to complete that particular run. It also works for partial
selections.

It is no longer necessary to call private/build-test-packages, ever.

The invocation of that script was dropped from our autopktest
declaration. Calling it there would not hurt, but it is not needed.

The file containing the script was not removed from our repo yet even
though it appears to be superfluous now. First, we have to determine
conclusively whether test packages should be packaged separately as
contemplated in Bug#926409 [1] although that currently appears to be
neither necessary nor a good technical solution.

[1] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=926409#42