Use original, unresolved paths when referring to and processing command line arguments.
For users, the paths are only visible in the experimental JSON output
format, but keeping the orginal paths is helpful when correlating that
output with command line arguments pointing to files in a local mirror
of the Debian archive.

Automated runners that generate packaging hints for the Lintian
website will be the primary beneficiaries of this change.