Turn maybe-not-arch-all-binnmuable into a classification tag.
The tag was introduced in an attempt to figure how many packages might
break if 'Arch: all' packages were binNMUed. [1] The idea was that the
versioned prerequisites between installables from the same source
would probably break.

The tag was marked experimental but should also be a classification
tag. This commit makes it so.

Thanks to Mattia Rizzolo for providing a valuable opinion and
historical context!

Thanks to Holger Levsen for bringing the matter to our attention!

[1] https://wiki.debian.org/binNMU