For odd-mark-in-description, provide line number for improved context.
The context did not indicate with any specificity how the offending
line may be located. We hope that a line number indicating the
problem's position relative to the description's beginning is an
acceptable remedy.

The package tuxmath, which h01ger mentioned on IRC, now triggers the
following hints (unrelated tags omitted). Please note the line
numbers:

$ bin/lintian /mirror/debian/pool/main/t/tuxmath/tuxmath_2.0.3-7_amd64.deb
P: tuxmath: odd-mark-in-description comma not followed by whitespace (line 1)
P: tuxmath: odd-mark-in-description comma not followed by whitespace (line 3)

Thanks to Holger Levsen for bringing the matter to our attention!