In the test suite, require that tags are tested via the check that owns them.
Some tags were only tested in the generalized tracking tests. That could lead
to confusion when researching the performance of a particular check.

The standard method for doing so is to run the test suite only for the affected
check, but those tests did not cover all affected tags!

They do now, with the exception of untested tags.

Gbp-Dch: ignore