Provide a generalized method to grant well-documented and isolated exemptions.
Some toolchains build packages that provoke the same hint in many
related packages. Often, the tooling is well-respected or has been
around for a long time. Either way, a maintainer can usually do
nothing about it. The hints are therefore perceived as false
positives.

Some teams will install overrides, while others approach Lintian with
requests for exemptions. Those were usually applied in the Lintian
check, namely at the point of detection. Unfortunately that cluttered
the code and sometimes introduced bugs.

This commit provides a generalized and well-documented way to grant
exemptions. Most important the exemptions, which are called "screens,"
are now separate and isolated pieces of code.

A hint that was masked by a screen is treated much like one with an
override, except that it is shown with the letter M (instead of O). An
automatic comment shows the name of the screen that generated the
mask.

Going forward, exemptions can be counted. They also carry enough
information to be reversed in an educated manner when the time is
right.

To aid with that effort, the screens provide the names and email
addresses of the petitioners, a summary of the reasoning for the
exemption, and further references (usually a bug number).