Recommend the newer php-markdown over libmarkdown-php when embedded. (Closes: #898867)
The older, "classic" version works as a plugin for WordPress and
Smarty but has not been supported since 2013. It may be removed from
the Debian archive at some point.

The newer version is used in the movim package in Debian unstable at
the time of writing, and may also be around longer.