Rename three tag files to new '*.tag' extension; rename fields.
Some of these must have crossed with the name changes when an older
branch was merged in commit ee125788.

The oversight was not noticed because the profiles still load '*.desc'
files for backward compatibility with any custom profiles users may
use.

The test script in the preceeding commit now flags such files in
Lintian's distributed files (but not in custom profiles).

Gbp-Dch: ignore