Integrate data file debhelper/dh_addons-manual into the source code.
Also fixes a bug. Like all other addons, the addon for the ADA
libraries uses an underscore in the name instead of a hyphen.

There was no way to generate the data file automatically (which would
have been a good reason to keep it around). Manual maintenance is just
as easily done in the Perl check and puts the data closer to the code
that uses it.

Gbp-Dch: ignore