Adjust to debhelper's new install location for systemd service files. (Closes: #992465)
This is the second of two commits restoring the Salsa CI pipeline.

Debhelper now produces packages that place systemd service files into
/usr/lib/systemd. Lintian now flags all such files when they are installed
elsewhere. As noted in the bug [1] that may be ahead of the curve. One Lintian
maintainer also believes it may accelerate the transition.

[1] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=992465#15

A note on the blocking bug:

The underlying debhelper change required the adjustment of many packaging hints
in the Lintian test suite and initially prompted a request for a debhelper
backport for bullseye. [2]

[2] https://bugs.debian.org/992711

The previous commit, however, separately required the installation of debhelper
from unstable to cut the next Lintian release, which will close over forty bugs.
(Divergence is minimal, because bullseye was released only two weeks ago.) The
backport is now less urgent but still needed going forward.

While the patches in the bug report did not cover the scope of this commit,
they made the work a lot easier. In recognition, authorship was therefore given
to Peter Pentchev.

Thank you, Peter, for those patches!