Allow failures in Salsa-CI step 'Lintian'
The job 'lintian' fails due to Lintian warnings as the pipeline has
SALSA_CI_LINTIAN_FAIL_WARNING since commit 3d388fbff.

The purpose of a CI system is to catch regressions. The normal state of a
CI pipeline should be green. The main branch of a software project should
ideally be protected and stay green forever as commits that fail in the CI
would never be allowed to be merged on on the main branch.

Using the 'allow_failure' mode will continue to show if the step fails
or passes but indicate to contributors that the step does not need to pass
for a new commit to be acceptable.

Seeing the Lintian project CI failing for a long time now hurts the
reputation of Lintian as a tool of high quality promoting the high quality
of other Debian software. With this change the Lintian Salsa-CI pipeline
will no longer be in a constantly failing mode.

This commit can be reverted the day the job 'lintian' is no longer failing.