Fix two failing test cases for autoconf 2.71 on Salsa CI in unstable.
For build-systems/automake/automake-general, the error message was:

  autoreconf: error: configure.in: AC_INIT not found; not an autoconf script?
  dh_autoreconf: error: autoreconf -f -i returned exit code 1

By the way, it is not clear that this test and the corresponding check belong
into a separate category for 'automake'. The naming of the configure file seems
to be a matter of autotools proper.

For
build-systems/autotools/cruft-autotools-pkg-config-macro-not-cross-compilation-safe,
the error message was

  autoreconf: warning: 'configure.ac' and 'configure.in' both present.
  autoreconf: proceeding with 'configure.ac'
  autoreconf: error: configure.ac: AC_INIT not found; not an autoconf script?
  dh_autoreconf: error: autoreconf -f -i returned exit code 1

That test was split into two: one for the common configure.ac filename and
another for the older configure.in name.

The missing AC_INIT declarations were added to all three tests. Also, all three
tests were marked to build only sources (the installables are not needed).

Gbp-Dch: ignore