In the test suite, rename expected 'tags' to the more accurately titled 'hints'; utilities too.
Renames all manner of things from tags* to hints* throughout the test suite. They
were left behind in the first big rename.

Follows the proposed convention that the hints intended for maintainers are
composed of tags that are augmented by a helpful context. The word 'tag' now
refers to the condition that gave rise to a hint's issuance.

Gbp-Dch: ignore