Fix autopkgtest failure on i386.
The tracker says "Not a regression" but that is not true:

    https://ci.debian.net/data/autopkgtest/testing/i386/l/lintian/15214685/log.gz

For some errors it is worthwhile to investigate how they occur. Here, I am not
sure why the previous code worked.

The proper way to access the exit status would have been $(.SHELLSTATUS). Now
we compare the string from the output instead. It is simpler and more elegant.

Gbp-Dch: ignore