Exempt empty packages so declared from empty-binary-package. (Closes: #994711)
For the 'gnat' package mentioned in the bug report, Lintian no longer produces
any hints whatsoever. The lone 'empty-binary-package' tag is gone:

    $ bin/lintian /mirror/debian/pool/main/g/gnat/gnat_10.1_all.deb

Unfortunately, Lintian's website can presently not receive updates. [1]

Provides a new, false-positive test case.

Modifies the standard template for d/control, which read in part "This may be
an empty package", so not all test cases are automatically exempt.

Thank you to Nicolas Boulenguez for bringing this matter to our attention, and
also for the patch!

[1] https://lists.debian.org/debian-lint-maint/2021/09/msg00148.html