Fix minimum version for init-system-helpers in a tag description. (Closes: #918137)
The bug report gave a few hints about version numbers that could be correct,
but did not pipoint one beyond the shadow of a doubt.

According to my research, the functionality was added in this commit [1] and
released in this one [2] as part of init-system-helpers 1.52. The changelog [3]
states so, too:

  init-system-helpers (1.52) unstable; urgency=medium

    [ Felipe Sateler ]
    * invoke-rc.d: add option to do nothing for native systemd units.
      It is useful to simplify maintainer scripts, since it allows executing
      this command for sysvinit/openrc
      systems, and deb-systemd-invoke for systemd systems

Thank you to Ferenc Wágner for bringing the matter to our attention!

Sorry you waited so long for a fix.

[1] https://salsa.debian.org/debian/init-system-helpers/-/commit/d45b6763760ec9ee3019e09af2e4a2ad33977f0c
[2] https://salsa.debian.org/debian/init-system-helpers/-/commit/5e68389bddd66b2dc2ee3f8296b24e041e0d1bf9
[3] https://sources.debian.org/src/init-system-helpers/1.60/debian/changelog/#L120-L123