Be ambivalent between /lib and /usr/lib for systemd service files. (See: #992465)
In its version 13.5.2, Debhelper today reverted [1] a recent change installing
systemd files in /usr/lib rather than in /lib. This commit is Lintian's
adjustment to the new circumstances.

Debhelper's original changes [2] prompted the filing of the bug above and then
Peter's fix thereto, but in the meantime a lengthy discussion on d-devel
resulted in a preliminary consensus that individual packages should not actually
move files between the /lib and /usr/lib locations on their own. [3]

With this change, Lintian should be ambivalent between the two locations for
systemd, meaning that we issue here no guidance asking maintainers to move.

Further clarification from the Technical Committee is expected.

[1] https://salsa.debian.org/debian/debhelper/-/commit/3083286de05f5c4e209e777edd5800b0cffa5941
[2] https://salsa.debian.org/debian/debhelper/-/commit/d70caa69c64b124e3611c967cfab93aef48346d8
[3] https://lists.debian.org/debian-devel/2021/07/msg00126.html