Do not expect files in sudoers.d to have standard file permissions. (Closes: #995261)
The bug was probably introduced when the 'files' check was broken up. Over
several days, the nesting depth of conditionals was reduced (as well as their
absolute number) to make the code easier to read. Perhaps it at least helped the
reporting party to locate the bug quickly.

Thanks to Axel Beckert for the detailed and helpful bug report, and also for
the patch!