Lintian::Relation: Fix implication relationships with :any qualifier
Truth table (cells indicate whether the dependency on the left implies
the dependency at the top):

               pkg          pkg:i386     pkg:any
    pkg        yes          maybe        yes
    pkg:i386   maybe        yes          yes
    pkg:any    no           no           yes

where the relationships labelled "maybe" are implied if the dependency
happens to be found in an i386 package, and otherwise not implied.

This fixes a false positive, #995490, which could be seen in the
scripts-control-interpreters-python test package: a dependency on python3
implies a dependency on python3:any, so it is enough to be able to run
a "#!/usr/bin/python3" script.

Closes: #995490
Fixes: 9bc560a6 "Implement the multiarch selector :any in package relationships"
Signed-off-by: Simon McVittie <smcv@debian.org>