interpreters: Allow a dependency on make:any for #!/usr/bin/make scripts
make:any is a weaker dependency than make, but it is sufficient to get
a /usr/bin/make executable that can be used for simple
architecture-agnostic build processes.

The make interpreter is not a completely architecture-independent
interface (see the changelog for versions 4.0-5 and 4.1-1) so it
cannot be Multi-Arch: foreign. A dependency on make:any declares that the
dependent package is only using the architecture-independent subset of
its interface, and therefore can use a foreign make package, as though
make was Multi-Arch: foreign.

Closes: #994902
Signed-off-by: Simon McVittie <smcv@debian.org>