Migrate 'setup' entry points in checks to lazy Moo attributes; merge 'breakdown' into $type.
The 'installable' entry point now always follows 'visit_control_files' and
'visit_installed_files' so it can be used instead of 'breakdown_X_files'.

Likewise, 'source' always follows 'visit_orig_files' and 'visit_patched_files',
but none of them are in this commit.

Lazy Moo attributes can easily replace the old 'setup_X_files' entry points.
One just has to be careful when issuing tags in the Moo attributes. It is
possible, but lazy attributes are often evaluated inside conditions. The tags
are then not always issued, and the surrounding conditionals are located far
away.

The last point triggered a partial rewrite of the systemd check, which can be
found in the previous commit.

Gbp-Dch: ignore