Make the condition of a missing interpreter prerequisite look less like a recommendation.
The purpose is to be transparent about the diagnostic circumstances and
therefore about the minimum condition that must be met by any remedy.

At she same time, we hope that the new output format deters users from adding
the :any multiarch acceptor to their prerequisites even if it appears in the
hint context. We do not issue a recommendation here but present a diagnosis.

Lintian's recommendation's can be found in the corresponding tag descriptions,
which we have already made an effort to adjust [1] although more adjustments may
be needed.

[1] https://salsa.debian.org/lintian/lintian/-/commit/4434d6594d9bccdbb6dfe9ab57cf709335d86eee