Clarify in tag descriptions for debian/rules check that hint context is not a recommendation. (Closes: #995981)
In the spirit of commit cd1c170c, modifies both the tag description and the
hint context for rules-require-build-prerequisite so the wording leaves no doubt
that the codition being diagnosed is not a recommendation.

In view of the importance of this issue, we then did something similar for the
nearby tag missing-build-depends-for-clean-target-in-debian-rules, which seemed
to be similarly affected.

In our view, there is no need to disable the tag merely based on what may be a
maintainer error for a single package. Please provide evidence that the problem
is widespread and not remedied by this commit. Perhaps additional sources so
affected will also cause FTCBFS, but the Lintian maintainers are not in the
loop there. Your help will be much appreciated!

As a side node, it possible that the check the reporting party was concerned
about was furthermore affected by Bug#995490 and Bug#995498. Both bugs lead to
widespread reporting of funny prerequisites. It resulted from the same
underlying condition that has since been fixed.

Updated hints for for the 'assimp' sources are available on our website. [1]

Thanks to Helmut Grohne for bringing the matter to our attention!

[1] https://lintian.debian.org/sources/assimp