Accept nodejs:any as a prerequisite for Node.js scripts. (Closes: #996710)
Lintian went through changes recently to accomodate the :any multiarch acceptor
in package relationships. For interpreters, a consequence is that the condition
being tested for---which should in many cases allow :any---is generally not the
same as the prerequisite being recommended, which never includes :any.

For now, the :any must be added explicitly for any interpreter for which it is
appropriate. That is done here for nodejs.

It is currently not clear whether the :any should be added to all interpreters.
Apparently, it does not make sense for some, such as bash.

The offending hint from the bug report is now gone:

➤ bin/lintian /mirror/debian/pool/main/a/acorn/node-acorn_8.5.0+ds+\~cs23.9.6-3_all.deb
W: node-acorn: package-relation-with-self Breaks: node-acorn (<< 8.5.0+ds+~cs23.9.6-2~)

Thanks to Yadd for bringing the matter to our attention!