Add new tag for bash-centric terms in regular (non-maintainer scripts.
Previously, re-used the same tag intended only for maintainer scripts.

In packages generated from the 'eye' sources, there were issues with \u0000
being present. For now, converts the match to hexadecimals when any characters
are present that are outside the ASCII (POSIX) printable set.

Also, changes the order of the context for better sorting of hints in EWI
output.

Gbp-Dch: ignore