Offer classification tag for generated source files. (Closes: #993662)
In the spirit of compromise, the author implemented the suggestion as a
classification tag. That is the lowest kind of scan result Lintian can offer.

The tag will not be shown to users of Lintian executable. It will also not
appear on the website (except the description). Interested parties can get the
results from the JSON interface of our website. [1]

The author is not sure the tag provides any benefit over the query on
codesearch.d.o suggested by Lintian co-maintainer Chris Lamb, but we try to
accomadate everyone.

The tag comes with two test cases.

The Lintian maintainers reserve the right to drop the tag if its performance
does not justify the perceived benefits.

At the same time, the tag could easily be elevated to a user-facing visibility.
Any change in status, however, would have to find broader support. The author
personally concurs with Sean Whitton.

Thanks to Bastien Roucariès for an early draft of this code!

[1] https://lintian.debian.org/query