Limit Python dot directories to those starting with a dot. (Closes: #997662)
Many steps have been taken to reduce the complexity of conditionals and their
nesting depth. The goal is to make code base easier to understand so users can
follow the changes. Here, a transcription error resulted in a bug.

This information

-package-contains-python-dot-directory                ~~ ^usr/lib/python[^\/]+/(dist|site)-packages/(|.+/)\.[^\/]+/$

-        $self->hint($tag, $file->name)
-          if $file->name =~ m/$regex/;
-    }

was translated into that code:

+    $self->hint('package-contains-python-dot-directory', $file->name)
+      if $file->dirname
+      =~ m{^ usr/lib/python[^/]+ / (?:dist|site)-packages / }x
+      && $file->name =~ m{ [.][^/]+ / }x;

but the regular expression was anchored improperly. The present commit says in
Lintian's arcane syntax that directories (the trailing slash) should start with
a dot (the leading slash).

The package mentioned in the bug report no longer produces any offending hints:

➤ bin/lintian /mirror/debian/pool/main/b/beautifulsoup4/python3-bs4_4.9.3-1_all.deb
[no hints]

Kudos for keeping the package Lintian-clean.

Thanks to Stefano Rivera for bringing the matter to our attention!