Multiarch terminology: Use "qualifier" instead of "acceptor"
Update tag descriptions of rules-require-build-prerequisite and
missing-build-depends-for-clean-target-in-debian-rules. Also rename
method multiarch_acceptor() to multiarch_qualifier() in
Lintian::Relation::Predicate.

Multiarch is hard enough as it is. Lets not confuse people with new
terminologies where there are already existing ones. What Lintian used
to call "acceptor" is a qualifier as documented originally here:

https://wiki.ubuntu.com/MultiarchCross