Accommodate dune's file names for the byte code in OCaml interfaces. (Closes: #999601)
This commit is untested but will hopefully cover all cases mentioned in the
bug, except the .private directory.

For further improvements, please send references to the installable packages in
the archive that trigger the the false positives you would like to avoid.

Thanks to Julien Puydt for bringing this matter to our attention!