Tolerate multiarch acceptors in prerequisites for Debhelper commands and addons. (Closes: #1000234)
This fixes the issue reported here and in the cloned Bug#999810, although for
the wrong reason. Lintian currently assumes that prerequisites are met when the
unrestricted declared prerequisites imply, or satisfy, a particular requirement.

In the present situation, Lintian replaces the :native multiarch acceptor on
python3-sphinx with :any (no restrictions!) and compares it with the weakest
assertion we can make about prerequisites, which is also :any. Hey, presto!

That logic is flawed and will be fixed in additional commits in the near future.

The offending hint in cmake is now gone even if the preceeding commit is
reverted. Please note the comment about that in the preceeding commit.

Thanks to Timo Röhling for bringing the matter to our attention!