Bump the maximum plausible bug number to 1,500,000. (Closes: #1000148)
The previous threshold of 1MM was reached earlier today.

As a side note, I wish the project would start to encode bug numbers in Base58
and use a prefix other than the hash used for comments (and now Salsa issues).
Too often, Git dropped bug numbers after mistaking them for comments.

Perhaps a bracketed asterisk like that [*68Jw] would work? It is the number of
the present bug in Base58.